Introduce HybridRepSetReader.cpp
